extract improved scan command patch
